requirements.txt
# 基础服务器库
Django==4.2.13
django_tinymce==3.6.1
docxtpl~=0.17.0
django-ninja~=1.2.2
django-ninja-jwt~=5.3.2
django-ninja-extra~=0.21.2
ninja_schema==0.13.6
numpy==1.24.2
pydantic~=2.7.4
python-docx==1.1.2
celery==5.3.1
mysqlclient~=2.2.4
user-agents~=2.2.0
django-environ~=0.11.2
Pillow~=10.4.0
# pandas解析html的表格用的库
pandas==2.0.3
html5lib==1.1
# 生成环境才使用库
waitress==3.0.0
# 暂时没用到写在项目里面
django-redis~=5.4.0
redis~=5.0.3
pyinstaller~=6.8.0
